Login functionality
Added the login functionality; takes in username and password and verifies with all the entries in the local array database and checks whether the user is registered or not. if not, then login fails and if the user exists (i.e. is registered already) then the user is logged in successfully.